moved the clustering process inside wdfml and single wavelet map

+++ wdfml/observers/SingleEventPrintFileObserver.py
@@ -6,36 +6,58 @@ __version__ = "1.0.0"
 __maintainer__ = "Elena Cuoco"
 __email__ = "elena.cuoco@ego-gw.it"
 __status__ = "Development"
-import logging
 
-from wdfml.observers.observer import Observer
 import csv
 import os.path
-from wdfml.structures.SingleEvent import *
-import numpy as np
+from wdfml.observers.observer import Observer
+
 
 class SingleEventPrintTriggers(Observer):
-    def __init__(self, par):
+    def __init__ ( self, par, fullPrint=0 ):
         self.filesave = par.dir + 'WDFTrigger-%s-GPS%s-AR%s-Win%s-Ov%s-SNR%s.csv' % (
-            par.chan, int(par.gpsStart), par.ARorder, par.window, par.overlap, int(par.threshold))
-        self.event = SingleEvent(par.window)
-        self.Ncoeff=par.window
-        try:
-            os.remove(self.filesave)
-        except OSError:
-            pass
+            par.chan, int(par.gps), par.ARorder, par.window, par.overlap, int(par.threshold))
+        self.id = 0
+        if os.path.isfile(self.filesave):
+            try:
+                os.remove(self.filesave)
+            except OSError:
+                pass
+        self.fullPrint = fullPrint
+        self.headers = ['gps', 'snr', 'freq', 'duration', 'wave']
+        self.headersFull = ['gps', 'snr', 'freq', 'duration', 'wave']
+        for i in range(par.Ncoeff):
+            self.headers.append("wt" + str(i))
+            self.headersFull.append("wt" + str(i))
+        for i in range(par.Ncoeff):
+            self.headersFull.append("rw" + str(i))
 
+            ### write on disk in ordered way
 
-    def update(self, ev):
+    def update ( self, CEV ):
         self.file_exists = os.path.isfile(self.filesave)
-        Cmax=np.zeros(self.Ncoeff)
-        for i in range(0, self.Ncoeff):
-            Cmax[i] = ev.GetCoeff(i)
-        self.event.update(ev.mTime, ev.mSNR, ev.mlevel, ev.mWave, Cmax)
-        self.EV = self.event.__dict__
-        with open(self.filesave, 'a') as csvfile:
-            headers = sorted(self.EV.keys())
-            writer = csv.DictWriter(csvfile, delimiter=',', lineterminator='\n', fieldnames=headers)
-            if not self.file_exists:
-                writer.writeheader()
-            writer.writerow(self.EV)
+        self.ev = CEV.__dict__
+        self.id += 1
+        self.ev['ID'] = self.id
+        if self.fullPrint == 0:
+            with open(self.filesave, 'a') as csvfile:
+                headers = ['gps', 'snr', 'freq', 'duration', 'wave']
+                toprint = dict((k, self.ev[k]) for k in
+                               ('gps', 'snr', 'freq', 'duration', 'wave'))
+                writer = csv.DictWriter(csvfile, delimiter=',', lineterminator='\n', fieldnames=headers)
+                if not self.file_exists:
+                    writer.writeheader()
+                writer.writerow(toprint)
+        if self.fullPrint == 1:
+            with open(self.filesave, 'a') as csvfile:
+                writer = csv.DictWriter(csvfile, delimiter=',', lineterminator='\n', fieldnames=self.headers)
+                toprint = dict((k, self.ev[k]) for k in self.headers)
+                if not self.file_exists:
+                    writer.writeheader()
+                writer.writerow(toprint)
+        if self.fullPrint == 2:
+            with open(self.filesave, 'a') as csvfile:
+                writer = csv.DictWriter(csvfile, delimiter=',', lineterminator='\n', fieldnames=self.headersFull)
+                toprint = dict((k, self.ev[k]) for k in self.headersFull)
+                if not self.file_exists:
+                    writer.writeheader()
+                writer.writerow(toprint)
